requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
pandas==2.2.1
plotly==5.19.0
python-telegram-bot==20.8
python-dotenv==1.0.1
flask-talisman==1.1.0 